Add WPF support
