+++ windows/RNSqlite2Package.cs
@@ -1,0 +1,32 @@
+using ReactNative.Bridge;
+using ReactNative.Modules.Core;
+using ReactNative.UIManager;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace RNSqlite2
+{
+    public class RNSqlite2Package : IReactPackage
+    {
+        public IReadOnlyList<INativeModule> CreateNativeModules(ReactContext reactContext)
+        {
+            return new List<INativeModule>
+            {
+                new RNSqlite2Module(reactContext)
+            };
+        }
+
+        public IReadOnlyList<Type> CreateJavaScriptModulesConfig()
+        {
+            return new List<Type>(0);
+        }
+
+        public IReadOnlyList<IViewManager> CreateViewManagers(ReactContext reactContext)
+        {
+            return new List<IViewManager>(0);
+        }
+    }
+}
